% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_scores.R
\name{normal_scores}
\alias{normal_scores}
\title{Calculating Normal Score and Standard(Z) Score}
\usage{
normal_scores(data, ties = "average")
}
\arguments{
\item{data}{Numeric vector. The standard score and normal score are calculated on this argument.}

\item{ties}{one of "average", "random", "first", and "last". The method to deal with the ties in data. Default is "average"}
}
\value{
data.frame with calculated normal score and standard score of the given data.
}
\description{
Calculating Normal Score and Standard(Z) Score
}
\examples{
data<-rnorm(30)
normal_scores(data = data, ties = "average")
}
