% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tertile.R
\name{tertile}
\alias{tertile}
\title{Calculate the 33rd and 66th percentiles of a vector}
\usage{
tertile(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
A named vector containing the 33rd and 66th percentiles of the input.
}
\description{
Calculate the 33rd and 66th percentiles of a vector
}
\examples{
quantile(c(1:100))

}
